pyproject.toml
[project]
name = "mini-instagram"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "celery[redis] (>=5.6.2,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
isort = "^7.0.0"
flake8 = "^7.0.0"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
